+++ CHANGES.rst
@@ -91,6 +91,7 @@ Development
 - Constructor refactoring and naming things
 - Add ``id`` attribute to bus configuration settings
 - Refactor and improve DS18B20 settings, reading and processing
+- Improve prettified sensor readings log output
 
 
 2019-06-22 0.5.1